fixed type error in tabular data module
