+++ api/src/datasets/datasets.service.ts
@@ -117,21 +117,6 @@ export class DatasetsService {
     }
     df = pl.readCSV(csvString, { tryParseDates: true });
 
-    // for datasets without primary keys, generate a sequential id column
-    if (createTabularDatasetDto.primaryKeys.length === 0) {
-      const indexArray = [];
-      for (let i = 0; i < df.shape.height; i++) {
-        indexArray.push(i);
-      }
-      const indexSeries = pl.Series('id', indexArray);
-      df.insertAtIdx(0, indexSeries);
-      createTabularDatasetDto.primaryKeys.push('id');
-    }
-
-    if (!this.primaryKeyCheck(createTabularDatasetDto.primaryKeys, df)) {
-      throw new ForbiddenException('Dataset failed primary keys check!');
-    }
-
     const dataset = await this.datasetModel.create({
       data: {
         datasetType: 'TABULAR',
@@ -142,139 +127,14 @@ export class DatasetsService {
       }
     });
 
-    const tabularData = await this.tabularDataModel.create({
-      data: {
-        datasetId: dataset.id,
-        primaryKeys: createTabularDatasetDto.primaryKeys
+    const tabularData = await this.tabularDataService.create(df, dataset.id, createTabularDatasetDto.primaryKeys);
+
+    return await this.datasetModel.update({
+      data: {},
+      where: {
+        id: dataset.id
       }
     });
-
-    for (let col of df.getColumns()) {
-      // create a float column
-      if (col.isFloat()) {
-        this.columnService.createColumn({
-          dataPermission: 'MANAGER',
-          floatData: col.toArray(),
-          kind: 'FLOAT',
-          name: col.name,
-          nullable: col.nullCount() != 0,
-          // numericColumnValidation: {
-          //   max: col.max(),
-          //   min: col.min()
-          // },
-          summary: {
-            count: col.len(),
-            max: col.max(),
-            mean: col.mean(),
-            median: col.median(),
-            min: col.min(),
-            notNullCount: col.len() - col.nullCount(),
-            // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-            std: col.rollingStd(col.len())[-1]
-          },
-          summaryPermission: 'MANAGER',
-          tabularDataId: tabularData.id
-        });
-        continue;
-      }
-
-      // create an int column
-      if (col.isNumeric()) {
-        this.columnService.createColumn({
-          dataPermission: 'MANAGER',
-          intData: col.toArray(),
-          kind: 'INT',
-          name: col.name,
-          // numericColumnValidation: {
-          //   max: col.max(),
-          //   min: col.min()
-          nullable: col.nullCount() != 0,
-          // },
-          summary: {
-            count: col.len(),
-            max: col.max(),
-            mean: col.mean(),
-            median: col.median(),
-            min: col.min(),
-            // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-            mode: col.mode()[0],
-            notNullCount: col.len() - col.nullCount(),
-            // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-            std: col.rollingStd(col.len())[-1]
-          },
-          summaryPermission: 'MANAGER',
-          tabularDataId: tabularData.id
-        });
-        continue;
-      }
-
-      // create a boolean column
-      if (col.isBoolean()) {
-        this.columnService.createColumn({
-          dataPermission: 'MANAGER',
-          enumData: col.toArray(),
-          kind: 'ENUM',
-          name: col.name,
-          nullable: col.nullCount() != 0,
-          summary: {
-            count: col.len(),
-            // distribute needs to be a record of string: number
-            distribution: {},
-            // col.valueCounts() returns a dataframe,
-            notNullCount: col.len() - col.nullCount()
-          },
-          summaryPermission: 'MANAGER',
-          tabularDataId: tabularData.id
-        });
-        continue;
-      }
-
-      // create a datetime column
-      if (col.isDateTime()) {
-        this.columnService.createColumn({
-          dataPermission: 'MANAGER',
-          // date is represented as time difference from 1970-Jan-01
-          // datetimeColumnValidation: {
-          //   max: new Date(),
-          //   min: '1970-01-01'
-          // },
-          // datetime is represented as milliseconds from 1970-Jan-01 00:00:00
-          datetimeData: col.toArray(),
-          kind: 'DATETIME',
-          name: col.name,
-          nullable: col.nullCount() != 0,
-          summary: {
-            count: col.len(),
-            max: new Date(),
-            min: new Date('1970-01-01'),
-            notNullCount: col.len() - col.nullCount()
-          },
-          summaryPermission: 'MANAGER',
-          tabularDataId: tabularData.id
-        });
-        continue;
-      }
-
-      // create a string column
-      this.columnService.createColumn({
-        dataPermission: 'MANAGER',
-        kind: 'STRING',
-        name: col.name,
-        // stringColumnValidation: {
-        //   minLength: 0
-        nullable: col.nullCount() != 0,
-        // },
-        stringData: col.toArray(),
-        summary: {
-          count: col.len(),
-          notNullCount: col.len() - col.nullCount()
-        },
-        summaryPermission: 'MANAGER',
-        tabularDataId: tabularData.id
-      });
-    }
-
-    return dataset;
   }
 
   async deleteDataset(datasetId: string, currentUserId: string) {
@@ -367,7 +227,7 @@ export class DatasetsService {
       // row range should be handled in the tabular data level
       // data type filter should be handled in the tabular data level
       // intra-column data hashing and trimming should be handled by the column service
-      const tabularDataView = await this.tabularDataService;
+      // const tabularDataView = await this.tabularDataService;
       return 'THE DATASET VIEW';
     }
 
@@ -423,20 +283,5 @@ export class DatasetsService {
       }
     });
     return await this.prisma.$transaction([updateDataset]);
-  }
-
-  private primaryKeyCheck(primaryKeys: string[], df: DataFrame): boolean {
-    for (let key of primaryKeys) {
-      const col = df.getColumn(key);
-      if (col.nullCount() > 0) {
-        return false;
-      }
-    }
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    const checkPrimaryKeysArray: boolean[] = df
-      .select(...primaryKeys)
-      .isUnique()
-      .toArray();
-    return checkPrimaryKeysArray.reduce((prev, curr) => prev && curr);
   }
 }
