+++ api/src/tabular-data/zod/tabular-data.ts
@@ -1,5 +1,7 @@
 import { z } from 'zod';
 
+import { $TabularColumn } from '@/columns/zod/columns';
+
 const $GetTabularDataViewDto = z.object({
   columnIds: z.string().array(),
   tabularDataId: z.string()
@@ -7,6 +9,15 @@ const $GetTabularDataViewDto = z.object({
 
 export type GetTabularDataViewDto = z.infer<typeof $GetTabularDataViewDto>;
 
-const $CreateTabularColumnDto = z.object({
-  datasetId: z.string()
+const $TabularData = z.object({
+  columns: $TabularColumn.array(),
+  datasetId: z.string(),
+  id: z.string(),
+  primaryKeys: z.string().array()
 });
+export type TabularData = z.infer<typeof $TabularData>;
+
+const $UpdatePrimaryKeysDto = z.object({
+  primaryKeys: z.string().array()
+});
+export type UpdatePrimaryKeysDto = z.infer<typeof $UpdatePrimaryKeysDto>;
